Fix missing self.artifacts in new filter
